html/browsers/browsing-the-web/back-forward-cache/dedicated-worker-broadcastchanel-evict-on-message.html

<!doctype html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
promise_test(async t => {
  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());
  const pageA2 = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;
  const urlA2 = location.origin + executorPath + pageA2.context_id;

  window.open(urlA, '_blank', 'noopener');
  await pageA.execute_script(waitForPageShow);

  await pageA.execute_script(() => {
    // Register a dedicated worker.
    let worker = new Worker('../resources/worker-with-broadcastchannel.js');
  });

  window.open(urlA2, '_blank', 'noopener');

  await pageA.execute_script(
    (url) => {
      prepareNavigation(() => {
        location.href = url;
      });
    },
    [urlB]
  );

  await pageB.execute_script(waitForPageShow);

  await pageA2.execute_script(() => {
    window.messages = [];
    let channel = new BroadcastChannel('foo');
    window.channel = channel;

    window.channel.addEventListener('message', event => {
      window.messages.push(event.data);
    });
    window.channel.postMessage('yay');
    window.channel.postMessage('yay');
  });

  let messages = await pageA2.execute_script(() => {
    return window.messages;
  });
  // Assert that the message has not arrived when PageA is in bfcache.
  assert_equals(messages.length, 0);

  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );
  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);

  let messagesAfterBfcache = await pageA2.execute_script(() => {
    return window.messages;
  });
  assert_equals(messagesAfterBfcache.length, 2);
}, 'Messages dispatched while in bfcache are delivered after being restored from bfcache.');
</script>
